cleaned up, added ButtonAnimatedSample
