+++ moonstone-extra/bower.json
@@ -3,7 +3,7 @@
   "dependencies": {
     "enyo": "git@github.com:enyojs/enyo.git#2.6.0-dev",
     "moonstone": "git@github.com:enyojs/moonstone.git#2.6.0-dev",
-    "moonstone-extra": "https://github.com/enyojs/moonstone-extra.git#master",
+    "moonstone-extra": "https://github.com/enyojs/moonstone-extra.git#76e549574685d65f2367d44546d2d29bf60d6ec2",
     "spotlight": "git@github.com:enyojs/spotlight.git#2.6.0-dev",
     "enyo-ilib": "git@github.com:enyojs/enyo-ilib.git#2.6.0-dev",
     "layout": "git@github.com:enyojs/layout.git#2.6.0-dev",
