+++ moonstone-extra/index.js
@@ -1,24 +1,8 @@
 var
-	strawman = require('../src');
+	strawman = require('./src');
 
 var 
-	ready = require('enyo/ready'),
-	kind = require('enyo/kind');
-
-var 
-
-	oldSampler = require('../src/moonstone-samples/lib/All');
-
-var
-	newSampler = kind({
-		kind: oldSampler
-	});
-
-newSampler.samples = {
-	ButtonSample: require('./src/ButtonSample.js')
-}
-
-strawman.samples['MoonstoneExtra'] = newSampler;
+	ready = require('enyo/ready');
 
 ready(function(){
 	var names = window.document.location.search.substring(1).split('&');
